move club/:club/ routes to new router
